+++ src/providers/Anthropic/index.ts
@@ -1,4 +1,4 @@
-import {Provider} from '../types';
+import {Provider} from '../../types';
 import {EmbeddingModel, generateText, LanguageModel} from 'ai';
 import {AnthropicProvider, createAnthropic} from '@ai-sdk/anthropic';
